Stop server's tickrate from drifting
