+++ client/CMakeLists.txt
@@ -9,11 +9,12 @@ target_sources(client PRIVATE
     "${CMAKE_CURRENT_LIST_DIR}/debug_overlay.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/deferred_pass.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/game.cpp"
-    "${CMAKE_CURRENT_LIST_DIR}/gamedata.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/gbuffer.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/globals.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/input.cpp"
-    "${CMAKE_CURRENT_LIST_DIR}/network_entities.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/net_ecs.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/net_entities.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/net_gamedata.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/screen.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/shadow_manager.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/shadowmap.cpp")
