+++ client/net_ecs.hpp
@@ -5,13 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 #pragma once
-#include <common/math/types.hpp>
-#include <entt/entt.hpp>
 
-namespace network_entities
+namespace net_ecs
 {
-void clear();
-entt::entity create(uint32_t network_id);
-entt::entity find(uint32_t network_id);
-void remove(uint32_t network_id);
-} // namespace network_entities
+void init();
+} // namespace net_ecs
