+++ client/net_gamedata.hpp
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2021, Kirill GPRB. All Rights Reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+#pragma once
+
+namespace net_gamedata
+{
+void init();
+} // namespace net_gamedata
