+++ shared/protocol/protocol.hpp
@@ -19,6 +19,7 @@ namespace protocol
 {
 constexpr static const uint16_t VERSION = 0x0001;
 constexpr static const uint16_t DEFAULT_PORT = 43103;
+constexpr static const float DEFAULT_TICKRATE = 30.0f;
 
 template<uint16_t packet_id>
 struct Packet { constexpr static const uint16_t id = packet_id; };
